+++ src/commonMain/kotlin/KSLog.kt
@@ -89,7 +89,6 @@ interface KSLog {
 
 operator fun KSLog.invoke(performLogCallback: (level: LogLevel, tag: String?, message: Any, throwable: Throwable?) -> Unit) = CallbackKSLog(performLogCallback)
 
-internal expect val defaultLogging: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit
 internal val printlnLogging: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit = { l, t, m, e ->
     println(defaultMessageFormatter(l, t, m, e))
 }
@@ -131,9 +130,9 @@ fun KSLog(
     levels: Iterable<LogLevel>,
     messageFormatter: MessageFormatter = defaultMessageFormatter
 ): KSLog {
-    val levels = levels.toSet()
+    val levelsSet = levels.toSet()
     return KSLog (defaultTag, messageFormatter).filtered { l, _, _ ->
-        l in levels
+        l in levelsSet
     }
 }