+++ src/mingwX64Main/kotlin/defaultLogging.kt
@@ -1,4 +1,0 @@
-package dev.inmo.kslog.common
-
-internal actual val defaultLogging: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit
-    get() = printlnLogging