+++ src/wasmJsMain/kotlin/KSLoggerDefaultPlatformLoggerLambda.kt
@@ -1,0 +1,4 @@
+package dev.inmo.kslog.common
+
+actual var KSLoggerDefaultPlatformLoggerLambda: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit =
+    printlnLogging